pyproject.toml
[tool.poetry]
name = "codellama-local-olama"
version = "0.1.0"
description = ""
authors = ["moosedev <ahmed.musawir@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.2.2"}
pydantic = ">=2.7.4,<3.0.0"
langchain-openai = "^0.1.8"
python-decouple = "^3.8"
langchain-anthropic = "^0.1.20"
langchain-google-genai = "^1.0.7"
langchain-groq = "^0.1.6"
langsmith = "^0.1.96"
langchain-community = "^0.2.15"
langchain = "^0.2.15"
langchain-core = "^0.2.37"
openai = "^1.43.0"
streamlit = "^1.38.0"
bs4 = "^0.0.2"
pillow = "^10.4.0"
langchain-ollama = "^0.1.3"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
